Part 2 Exc.2.5 separate module
