+++ part2/courses-contents/src/index.js
@@ -1,51 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-
-const Header = (props) => {
-  return (
-    <h1>Web development curriculum</h1>
-  )
-}
-
-const CourseName = (props) => {
-  return (
-    <h2>{props.name}</h2>
-  )
-}
-
-const Part = (props) => {
-  return (
-    <p>
-      {props.name} {props.exercises}
-    </p>
-  )
-}
-
-const Content = (props) => {
-  return props.parts.map(
-    part => <Part key={part.id} name={part.name} exercises={part.exercises}/>
-  )
-}
-
-const Total = (props) => {
-  const sum = props.parts.reduce( (sum, part) => {
-    return sum + part.exercises
-  }, 0)
-
-  return (
-    <strong>Total of {sum}</strong>
-  )
-}
-
-const Course = (props) => {
-  return props.courses.map(course =>
-    <div key={course.id}>
-      <CourseName name={course.name}/>
-      <Content parts={course.parts}/>
-      <Total parts={course.parts}/>
-    </div>
-  )
-}
+import Courses from './components/courses';
 
 const App = () => {
   const courses = [
@@ -95,8 +50,8 @@ const App = () => {
 
   return (
     <div>
-      <Header />
-      <Course courses={courses} />
+      <h1>Web development curriculum</h1>
+      <Courses courses={courses} />
     </div>
   )
 }
